# swagger-docs/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Book Catalogue API",
    "description": "This API handles operations related to books in the catalog.",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/books/isbn/{isbn}": {
      "get": {
        "tags": [
          "Books"
        ],
        "description": "Endpoint to search for a book by its ISBN.",
        "parameters": [
          {
            "$ref": "#/parameters/isbnPath"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "$ref": "#/responses/BookFound"
          },
          "400": {
            "description": "Bad Request",
            "$ref": "#/responses/InvalidISBN"
          },
          "404": {
            "description": "Not Found",
            "$ref": "#/responses/BookNotFound"
          },
          "500": {
            "description": "Internal Server Error",
            "$ref": "#/responses/ServerError"
          }
        }
      }
    },
    "/api/v1/books/": {
      "get": {
        "tags": [
          "Books"
        ],
        "description": "Endpoint to search for books with optional filters. You can filter by title, author, publication year, category, language, and featured type.",
        "parameters": [
          {
            "$ref": "#/parameters/titleQuery"
          },
          {
            "$ref": "#/parameters/authorQuery"
          },
          {
            "$ref": "#/parameters/publicationYearQuery"
          },
          {
            "$ref": "#/parameters/categoryQuery"
          },
          {
            "$ref": "#/parameters/languageQuery"
          },
          {
            "$ref": "#/parameters/featuredTypeQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "$ref": "#/responses/BooksFound"
          },
          "400": {
            "description": "Bad Request",
            "$ref": "#/responses/InvalidQueryParams"
          },
          "404": {
            "description": "Not Found",
            "$ref": "#/responses/NoBooksFound"
          },
          "500": {
            "description": "Internal Server Error",
            "$ref": "#/responses/ServerError"
          }
        }
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "description": "Create a new book in the catalog.",
        "parameters": [
          {
            "name": "bookBody",
            "in": "body",
            "description": "Book details to create a new book.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Book"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "$ref": "#/responses/BookCreated"
          },
          "400": {
            "description": "Bad Request",
            "$ref": "#/responses/ValidationError"
          },
          "409": {
            "description": "Conflict",
            "$ref": "#/responses/DuplicateISBN"
          },
          "500": {
            "description": "Internal Server Error",
            "$ref": "#/responses/ServerError"
          }
        }
      }
    },
    "/api/v1/books/featured": {
      "get": {
        "tags": [
          "Books"
        ],
        "description": "Endpoint to fetch all featured books, where the featured type is not \"none\".",
        "responses": {
          "200": {
            "description": "OK",
            "$ref": "#/responses/BooksFound"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found",
            "$ref": "#/responses/NoFeaturedBooks"
          },
          "500": {
            "description": "Internal Server Error",
            "$ref": "#/responses/ServerError"
          }
        }
      }
    },
    "/api/v1/books/latest": {
      "get": {
        "tags": [
          "Books"
        ],
        "description": "Endpoint to fetch the 10 latest books, ordered by publication year in descending order.",
        "responses": {
          "200": {
            "description": "OK",
            "$ref": "#/responses/BooksFound"
          },
          "404": {
            "description": "Not Found",
            "$ref": "#/responses/NoBooksFound"
          },
          "500": {
            "description": "Internal Server Error",
            "$ref": "#/responses/ServerError"
          }
        }
      }
    },
    "/api/v1/books/stats": {
      "get": {
        "tags": [
          "Books"
        ],
        "description": "Endpoint to fetch statistics about the books collection, including total books, number of authors, most popular genre, and most prolific author.",
        "responses": {
          "200": {
            "description": "OK",
            "$ref": "#/responses/StatsFetched"
          },
          "500": {
            "description": "Internal Server Error",
            "$ref": "#/responses/ServerError"
          }
        }
      }
    },
    "/api/v1/books/{isbn}/downloads": {
      "patch": {
        "tags": [
          "Books"
        ],
        "description": "Endpoint to update the download count for a specific book identified by its ISBN.",
        "parameters": [
          {
            "$ref": "#/parameters/isbnPath"
          },
          {
            "name": "downloadCountBody",
            "in": "body",
            "description": "Download count to update for the book.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "downloadCount": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "integer"
                        },
                        "example": {
                          "type": "number",
                          "example": 100
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "$ref": "#/responses/DownloadUpdated"
          },
          "400": {
            "description": "Bad Request",
            "$ref": "#/responses/ValidationError"
          },
          "404": {
            "description": "Not Found",
            "$ref": "#/responses/BookNotFound"
          },
          "500": {
            "description": "Internal Server Error",
            "$ref": "#/responses/ServerError"
          }
        }
      }
    },
    "/api/v1/books/{isbn}/readingLists": {
      "patch": {
        "tags": [
          "Books"
        ],
        "description": "Update the reading lists count for a book identified by its ISBN.",
        "parameters": [
          {
            "$ref": "#/parameters/isbnPath"
          },
          {
            "name": "readingListsBody",
            "in": "body",
            "description": "Reading lists count to update for the book.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "inReadingLists": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "integer"
                        },
                        "example": {
                          "type": "number",
                          "example": 50
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "$ref": "#/responses/inReadingListsUpdated"
          },
          "400": {
            "description": "Bad Request",
            "$ref": "#/responses/ValidationError"
          },
          "404": {
            "description": "Not Found",
            "$ref": "#/responses/BookNotFound"
          },
          "500": {
            "description": "Internal Server Error",
            "$ref": "#/responses/ServerError"
          }
        }
      }
    },
    "/api/v1/books/{isbn}/review": {
      "patch": {
        "tags": [
          "Books"
        ],
        "description": "Submit a review and update the total rating and review count.",
        "parameters": [
          {
            "$ref": "#/parameters/isbnPath"
          },
          {
            "name": "reviewBody",
            "in": "body",
            "description": "Review details to be submitted, including the score.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "score": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "number"
                        },
                        "example": {
                          "type": "number",
                          "example": 4
                        },
                        "description": {
                          "type": "string",
                          "example": "The score given to the book (0 to 5)."
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "$ref": "#/responses/ReviewAdded"
          },
          "400": {
            "description": "Bad Request",
            "$ref": "#/responses/ValidationError"
          },
          "404": {
            "description": "Not Found",
            "$ref": "#/responses/BookNotFound"
          },
          "500": {
            "description": "Internal Server Error",
            "$ref": "#/responses/ServerError"
          }
        }
      }
    },
    "/api/v1/books/{isbn}": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "description": "Delete a book from the catalog by ISBN.",
        "parameters": [
          {
            "$ref": "#/parameters/isbnPath"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "$ref": "#/responses/BookDeleted"
          },
          "400": {
            "description": "Bad Request",
            "$ref": "#/responses/InvalidISBN"
          },
          "404": {
            "description": "Not Found",
            "$ref": "#/responses/BookNotFound"
          },
          "500": {
            "description": "Internal Server Error",
            "$ref": "#/responses/ServerError"
          }
        }
      },
      "put": {
        "tags": [
          "Admin"
        ],
        "description": "Endpoint to update the details of a book by ISBN.",
        "parameters": [
          {
            "$ref": "#/parameters/isbnPath"
          },
          {
            "name": "bookBody",
            "in": "body",
            "description": "Updated details for the book.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Book"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "$ref": "#/responses/BookUpdated"
          },
          "400": {
            "description": "Bad Request",
            "$ref": "#/responses/InvalidISBN"
          },
          "404": {
            "description": "Not Found",
            "$ref": "#/responses/BookNotFound"
          },
          "500": {
            "description": "Internal Server Error",
            "$ref": "#/responses/ServerError"
          }
        }
      }
    }
  },
...
  "parameters": {
    "isbnPath": {
      "name": "isbn",
      "in": "path",
      "description": "ISBN of the book. Must be in ISBN-10 or ISBN-13 format.",
      "required": true,
      "type": "string",
      "format": "isbn"
    },
    "titleQuery": {
      "name": "title",
      "in": "query",
      "required": false,
      "type": "string",
      "example": "El día que se perdió la cordura"
    },
    "authorQuery": {
      "name": "author",
      "in": "query",
      "required": false,
      "type": "string",
      "example": "Javier Castillo"
    },
    "publicationYearQuery": {
      "name": "publicationYear",
      "in": "query",
      "required": false,
      "type": "integer",
      "example": 2015
    },
    "categoryQuery": {
      "name": "category",
      "in": "query",
      "required": false,
      "type": "string",
      "example": "Fiction"
    },
    "languageQuery": {
      "name": "language",
      "in": "query",
      "required": false,
      "type": "string",
      "example": "es"
    },
    "featuredTypeQuery": {
      "name": "featuredType",
      "in": "query",
      "required": false,
      "type": "string",
      "example": "bestSeller"
    }
  },
  "responses": {
    "BookFound": {
      "description": "Book found.",
      "schema": {
        "$ref": "#/definitions/Book"
      }
    },
    "BooksFound": {
      "description": "Books found.",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/Book"
        }
      }
    },
    "BookNotFound": {
      "description": "Book not found.",
      "schema": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Book not found"
          }
        }
      }
    },
    "StatsFetched": {
      "description": "Statistics fetched successfully.",
      "schema": {
        "$ref": "#/definitions/Stats"
      }
    },
    "DownloadUpdated": {
      "description": "Download count updated successfully.",
      "schema": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Book download count updated successfully."
          },
          "updatedBook": {
            "$ref": "#/definitions/Book"
          }
        }
      }
    },
    "inReadingListsUpdated": {
      "description": "Reading lists count updated successfully.",
      "schema": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Book total reading lists updated successfully."
          },
          "book": {
            "$ref": "#/definitions/Book"
          }
        }
      }
    },
    "ReviewAdded": {
      "description": "Review added successfully, returning the updated rating and review count.",
      "schema": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Review added successfully."
          },
          "bookReview": {
            "type": "object",
            "properties": {
              "isbn": {
                "type": "string",
                "example": "9788466346122"
              },
              "totalRating": {
                "type": "number",
                "example": 4.5
              },
              "totalReviews": {
                "type": "integer",
                "example": 10
              }
            }
          }
        }
      }
    },
    "BookCreated": {
      "description": "Book created successfully.",
      "schema": {
        "$ref": "#/definitions/Book"
      }
    },
    "BookDeleted": {
      "description": "Book deleted successfully.",
      "schema": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Book deleted successfully"
          }
        }
      }
    },
    "BookUpdated": {
      "description": "Book updated successfully.",
      "schema": {
        "$ref": "#/definitions/Book"
      }
    },
    "NoBooksFound": {
      "description": "No books found for the given search criteria.",
      "schema": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "No books found for the given search criteria."
          }
        }
      }
    },
    "NoFeaturedBooks": {
      "description": "No featured books found.",
      "schema": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "No featured books found."
          }
        }
      }
    },
    "InvalidISBN": {
      "description": "Invalid ISBN format. Must be ISBN-10 or ISBN-13.",
      "schema": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Invalid ISBN format. Must be ISBN-10 or ISBN-13."
          }
        }
      }
    },
    "DuplicateISBN": {
      "description": "Conflict error. Duplicate ISBN.",
      "schema": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Duplicate ISBN: a book with this ISBN already exists."
          }
        }
      }
    },
    "InvalidQueryParams": {
      "description": "Invalid query parameters provided.",
      "schema": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Invalid query parameters provided."
          },
          "invalidParameters": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "invalidParam1",
              "invalidParam2"
            ]
          }
        }
      }
    },
    "ValidationError": {
      "description": "Validation failed. Check the provided data.",
      "schema": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Validation failed. Check the provided data."
          },
          "details": {
            "type": "object",
            "example": {
              "isbn": "Invalid ISBN format"
            }
          }
        }
      }
    },
    "ServerError": {
      "description": "Unexpected server error occurred.",
      "schema": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Unexpected server error occurred."
          },
          "error": {
            "type": "string",
            "example": "InternalServerError: database connection failed."
          }
        }
      }
    }
  }
}